refactor: split out webpack configs
