+++ webpack/webpack-base.config.js
@@ -1,0 +1,32 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const webpack = require('webpack');
+
+const config = {
+  entry: './src/index.ts',
+  output: {
+    filename: 'index.js',
+    path: __dirname + '/dist',
+  },
+  devtool: 'source-map',
+  resolve: {
+    extensions: ['.ts', '.tsx', '.js', '.json'],
+  },
+  module: {
+    rules: [
+      { test: /\.tsx?$/, loader: 'awesome-typescript-loader' },
+      { enforce: 'pre', test: /\.js$/, loader: 'source-map-loader' },
+    ],
+  },
+  externals: {
+    react: 'React',
+    'react-dom': 'ReactDOM',
+  },
+  plugins: [
+    new webpack.DefinePlugin({
+      'process.env.NODE_ENV': JSON.stringify(process.env.NODE_ENV),
+    }),
+  ],
+};
+
+module.exports = config;
